+++ src/CommonITILCost.php
@@ -220,13 +220,13 @@ abstract class CommonITILCost extends CommonDBChild
             'joinparams'         => [
                 'jointype'           => 'child'
             ],
-            'computation'        =>
-            '(SUM(' . $DB->quoteName('TABLE.actiontime') . ' * ' .
-            $DB->quoteName('TABLE.cost_time') . '/' . HOUR_TIMESTAMP .
-            ' + ' . $DB->quoteName('TABLE.cost_fixed') . ' + ' .
-            $DB->quoteName('TABLE.cost_material') . ') / COUNT(' .
-            $DB->quoteName('TABLE.id') . ')) * COUNT(DISTINCT ' .
-            $DB->quoteName('TABLE.id') . ')',
+            'computation'        => new QueryExpression(
+                '(' . QueryFunction::sum(
+                    expression: $DB::quoteName('TABLE.actiontime') . ' * ' . $DB::quoteName('TABLE.cost_time') .' / ' .
+                        HOUR_TIMESTAMP . ' + ' . $DB::quoteName('TABLE.cost_fixed') . ' + ' . $DB::quoteName('TABLE.cost_material')
+                ) . ' / ' . QueryFunction::count($DB::quoteName('TABLE.id')) . ') * ' .
+                    QueryFunction::count(expression: $DB::quoteName('TABLE.id'), distinct: true)
+            ),
             'nometa'             => true, // cannot GROUP_CONCAT a SUM
         ];
 
@@ -242,11 +242,13 @@ abstract class CommonITILCost extends CommonDBChild
             'joinparams'         => [
                 'jointype'           => 'child'
             ],
-            'computation'        =>
-            '(SUM(' . $DB->quoteName('TABLE.actiontime') . ' * ' .
-            $DB->quoteName('TABLE.cost_time') . '/' . HOUR_TIMESTAMP .
-            ') / COUNT(' . $DB->quoteName('TABLE.id') . ')) * COUNT(DISTINCT ' .
-            $DB->quoteName('TABLE.id') . ')',
+            'computation'        => new QueryExpression(
+                '(' . QueryFunction::sum(
+                    expression: $DB::quoteName('TABLE.actiontime') . ' * ' . $DB::quoteName('TABLE.cost_time') .' / ' .
+                        HOUR_TIMESTAMP
+                ) . ' / ' . QueryFunction::count($DB::quoteName('TABLE.id')) . ') * ' .
+                    QueryFunction::count(expression: $DB::quoteName('TABLE.id'), distinct: true)
+            ),
             'nometa'             => true, // cannot GROUP_CONCAT a SUM
         ];
 
@@ -276,10 +278,12 @@ abstract class CommonITILCost extends CommonDBChild
             'joinparams'         => [
                 'jointype'           => 'child'
             ],
-            'computation'        =>
-            '(SUM(' . $DB->quoteName('TABLE.cost_fixed') . ') / COUNT(' .
-            $DB->quoteName('TABLE.id') . '))
-            * COUNT(DISTINCT ' . $DB->quoteName('TABLE.id') . ')',
+            'computation'        => new QueryExpression(
+                '(' . QueryFunction::sum(
+                    expression: $DB::quoteName('TABLE.cost_fixed')
+                ) . ' / ' . QueryFunction::count($DB::quoteName('TABLE.id')) . ') * ' .
+                    QueryFunction::count(expression: $DB::quoteName('TABLE.id'), distinct: true)
+            ),
             'nometa'             => true, // cannot GROUP_CONCAT a SUM
         ];
 
@@ -295,10 +299,12 @@ abstract class CommonITILCost extends CommonDBChild
             'joinparams'         => [
                 'jointype'           => 'child'
             ],
-            'computation'        =>
-            '(SUM(' . $DB->quoteName('TABLE.cost_material') . ') / COUNT(' .
-            $DB->quoteName('TABLE.id') . '))
-            * COUNT(DISTINCT ' . $DB->quoteName('TABLE.id') . ')',
+            'computation'        => new QueryExpression(
+                '(' . QueryFunction::sum(
+                    expression: $DB::quoteName('TABLE.cost_material')
+                ) . ' / ' . QueryFunction::count($DB::quoteName('TABLE.id')) . ') * ' .
+                    QueryFunction::count(expression: $DB::quoteName('TABLE.id'), distinct: true)
+            ),
             'nometa'             => true, // cannot GROUP_CONCAT a SUM
         ];