+++ src/Project.php
@@ -791,7 +791,7 @@ class Project extends CommonDBTM implements ExtraVisibilityCriteria
                     ],
                 ],
             ],
-            'computation'        => '(SUM(' . $DB->quoteName('TABLE.cost') . '))',
+            'computation'        => QueryFunction::sum($DB::quoteName('TABLE.cost')),
             'nometa'             => true, // cannot GROUP_CONCAT a SUM
         ];